Rough impl of learning loop
