+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod activation;
+pub mod layer;
 pub mod network;
 
 #[macro_export]
